requirements.txt
fastapi==0.109.0
supabase==2.3.4
uvicorn==0.27.0
python-decouple==3.8